[AdvancedBlacklist] Start work on modal
Renamed `menus.py` => `utils.py` and moved some methods to the file
