+++ advancedblacklist/patching.py
@@ -8,12 +8,11 @@
 
 import asyncio
 import logging
-from typing import TYPE_CHECKING, Any, Callable, Coroutine, Dict, Final, List
 from functools import wraps
+from typing import TYPE_CHECKING, Any, Callable, Coroutine, Dict, Final, List
 
 from discord.utils import maybe_coroutine
 from redbot.core.bot import Red
-
 
 __all__ = ["Patch"]
